backend/requirements.txt
aniso8601==9.0.1
bcrypt==4.3.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.18
dnspython==2.7.0
Farama-Notifications==0.0.4
filelock==3.18.0
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-PyMongo==3.0.1
Flask-RESTful==0.3.10
fsspec==2025.3.2
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
gymnasium==1.0.0
httplib2==0.22.0
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
limits==4.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
mysql-connector-python==9.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
redis==6.1.0
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
resend==2.6.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
